Split the pad wakeup allocation per ext1 version and index the pad table by pin number
